+++ src/test/scala/edu/ie3/simona/model/grid/GridSpec.scala
@@ -7,13 +7,15 @@
 package edu.ie3.simona.model.grid
 
 import java.util.UUID
-
 import breeze.linalg.DenseMatrix
 import breeze.math.Complex
 import breeze.numerics.abs
 import edu.ie3.datamodel.exceptions.InvalidGridException
+import edu.ie3.datamodel.models.input.MeasurementUnitInput
+import edu.ie3.datamodel.models.voltagelevels.GermanVoltageLevelUtils
 import edu.ie3.simona.exceptions.GridInconsistencyException
-import edu.ie3.simona.model.grid.GridModel.GridComponents
+import edu.ie3.simona.model.control.TransformerControlGroup
+import edu.ie3.simona.model.grid.GridModel.{GridComponents, GridControls}
 import edu.ie3.simona.test.common.input.{GridInputTestData, LineInputTestData}
 import edu.ie3.simona.test.common.model.grid.{
   BasicGrid,
@@ -21,6 +23,9 @@ import edu.ie3.simona.test.common.model.grid.{
   FiveLinesWithNodes
 }
 import edu.ie3.simona.test.common.{DefaultTestData, UnitSpec}
+import testutils.TestObjectFactory
+
+import scala.jdk.CollectionConverters.SetHasAsJava
 
 class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
 
@@ -117,7 +122,8 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
           Set(transformer2wModel),
           Set.empty[Transformer3wModel],
           switches
-        )
+        ),
+        GridControls(Set.empty[TransformerControlGroup])
       )
       // get the private method for validation
       val validateConnectivity: PrivateMethod[Unit] =
@@ -145,7 +151,8 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
           Set(transformer2wModel),
           Set.empty[Transformer3wModel],
           Set.empty[SwitchModel]
-        )
+        ),
+        GridControls(Set.empty[TransformerControlGroup])
       )
 
       // get the private method for validation
@@ -190,7 +197,8 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
           Set(transformer2wModel),
           Set.empty[Transformer3wModel],
           switches
-        )
+        ),
+        GridControls(Set.empty[TransformerControlGroup])
       )
 
       // get the private method for validation
@@ -229,7 +237,8 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
             Set(transformer2wModel),
             Set.empty[Transformer3wModel],
             switches
-          )
+          ),
+          GridControls(Set.empty[TransformerControlGroup])
         )
 
         // update the uuidToIndexMap
@@ -272,7 +281,8 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
             Set(transformer2wModel),
             Set.empty[Transformer3wModel],
             switches
-          )
+          ),
+          GridControls(Set.empty[TransformerControlGroup])
         )
 
         // update the uuidToIndexMap
@@ -357,7 +367,8 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
             Set(transformer2wModel),
             Set.empty[Transformer3wModel],
             switches
-          )
+          ),
+          GridControls(Set.empty[TransformerControlGroup])
         )
 
         // update the uuidToIndexMap
@@ -425,7 +436,8 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
             Set(transformer2wModel),
             Set.empty[Transformer3wModel],
             Set.empty[SwitchModel]
-          )
+          ),
+          GridControls(Set.empty[TransformerControlGroup])
         )
 
         // update the uuidToIndexMap
@@ -446,9 +458,88 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
         gridModel.nodeUuidToIndexMap.keySet.toVector.sorted should be(
           nodes.map(node => node.uuid).toVector.sorted
         )
-
       }
+    }
 
+    "build correct transformer control models" should {
+      /* Testing of distinct transformer control group building can be found in the spec for transformer control groups */
+
+      "determine node uuids correctly" in {
+        val determineNodeUuids =
+          PrivateMethod[Set[UUID]](Symbol("determineNodeUuids"))
+
+        val node0 = TestObjectFactory.buildNodeInput(
+          false,
+          GermanVoltageLevelUtils.MV_10KV,
+          1
+        )
+        val node1 = TestObjectFactory.buildNodeInput(
+          false,
+          GermanVoltageLevelUtils.MV_10KV,
+          1
+        )
+        val node2 = TestObjectFactory.buildNodeInput(
+          false,
+          GermanVoltageLevelUtils.MV_10KV,
+          1
+        )
+        val node3 = TestObjectFactory.buildNodeInput(
+          false,
+          GermanVoltageLevelUtils.MV_10KV,
+          1
+        )
+
+        val measurementUnits = Set(
+          new MeasurementUnitInput(
+            UUID.fromString("3ad9e076-c02b-4cf9-8720-18e2bb541ede"),
+            "measurement_unit_0",
+            node0,
+            true,
+            false,
+            false,
+            false
+          ),
+          new MeasurementUnitInput(
+            UUID.fromString("ab66fbb0-ece1-44b9-9341-86a884233ec4"),
+            "measurement_unit_1",
+            node1,
+            true,
+            false,
+            false,
+            false
+          ),
+          new MeasurementUnitInput(
+            UUID.fromString("93b4d0d8-cc67-41f5-9d5c-1cd6dbb2e70d"),
+            "measurement_unit_2",
+            node2,
+            true,
+            false,
+            false,
+            false
+          ),
+          new MeasurementUnitInput(
+            UUID.fromString("8e84eb8a-2940-4900-b0ce-0eeb6bca8bae"),
+            "measurement_unit_3",
+            node3,
+            false,
+            false,
+            false,
+            false
+          )
+        ).asJava
+        val selectedMeasurements = List(
+          "ab66fbb0-ece1-44b9-9341-86a884233ec4",
+          "93b4d0d8-cc67-41f5-9d5c-1cd6dbb2e70d",
+          "8e84eb8a-2940-4900-b0ce-0eeb6bca8bae"
+        )
+        val expectedUuids = Set(node1, node2).map(_.getUuid)
+
+        val actual = GridModel invokePrivate determineNodeUuids(
+          measurementUnits,
+          selectedMeasurements
+        )
+        actual should contain theSameElementsAs expectedUuids
+      }
     }
 
     "process a valid GridInputModel without an Exception" in new GridInputTestData {
@@ -456,10 +547,9 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
         validTestGridInputModel,
         gridInputModelTestDataRefSystem,
         defaultSimulationStart,
-        defaultSimulationEnd
+        defaultSimulationEnd,
+        controlConfig = None
       )
-
     }
   }
-
 }